Refactor into redux
