+++ src/app/store.js
@@ -1,7 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
+import problemsReducer from "../features/ourDesign/problemsSlice";
 
 const store = configureStore({
-  reducer: {},
+  reducer: {
+    problems: problemsReducer,
+  },
 });
 
 export default store;
