+++ src/views/Homepage.jsx
@@ -3,7 +3,7 @@ import styles from "./Homepage.module.css";
 
 import SplashImage from "../components/splashImage/SplashImage";
 import OurMission from "../components/ourMission/OurMission";
-import OurDesign from "../components/ourDesign/OurDesign";
+import OurDesign from "../features/ourDesign/OurDesign";
 
 const Homepage = () => {
   return (
